+++ Siv3D/src/Siv3D-Platform/macOS_Linux/Siv3D/Keyboard/CKeyboard.cpp
@@ -288,7 +288,7 @@ namespace s3d
 	//
 	////////////////////////////////////////////////////////////////
 
-	InputState& CKeyboard::getInputState(const size_t index) noexcept
+	InputState& CKeyboard::getInputState(const uint32 index) noexcept
 	{
 		assert(index < Keyboard::NumKeys);
 		return m_states[index];
